Consolidate crates and constants
