+++ monad-consensus-state/Cargo.toml
@@ -23,7 +23,6 @@ tracing = { workspace = true }
 
 [dev-dependencies]
 monad-eth-testutil = { path = "../monad-eth-testutil" }
-monad-eth-tx = { path = "../monad-eth-tx" }
 monad-eth-txpool = { path = "../monad-eth-txpool" }
 monad-eth-block-policy = { path = "../monad-eth-block-policy" }
 monad-eth-block-validator = { path = "../monad-eth-block-validator" }
