+++ monad-eth-block-policy/Cargo.toml
@@ -9,10 +9,10 @@ bench = false
 [dependencies]
 monad-consensus-types = { path = "../monad-consensus-types" }
 monad-crypto = { path = "../monad-crypto" }
-monad-eth-tx = { path = "../monad-eth-tx" }
-monad-state-backend = { path = "../monad-state-backend" }
 monad-eth-types = { path = "../monad-eth-types" }
+monad-state-backend = { path = "../monad-state-backend" }
 monad-types = { path = "../monad-types" }
+
 alloy-primitives = { workspace = true }
 reth-primitives = { workspace = true }
 itertools = { workspace = true }
