+++ monad-eth-ledger/Cargo.toml
@@ -13,7 +13,6 @@ monad-blocksync = { path = "../monad-blocksync" }
 monad-consensus-types = { version = "0.1.0", path = "../monad-consensus-types" }
 monad-crypto = { version = "0.1.0", path = "../monad-crypto" }
 monad-state-backend = { version = "0.1.0", path = "../monad-state-backend" }
-monad-eth-tx = { version = "0.1.0", path = "../monad-eth-tx" }
 monad-eth-types = { version = "0.1.0", path = "../monad-eth-types" }
 monad-executor = { version = "0.1.0", path = "../monad-executor" }
 monad-executor-glue = { version = "0.1.0", path = "../monad-executor-glue" }
