+++ monad-eth-ledger/src/lib.rs
@@ -20,8 +20,7 @@ use monad_consensus_types::{
 use monad_crypto::certificate_signature::{
     CertificateSignaturePubKey, CertificateSignatureRecoverable,
 };
-use monad_eth_tx::EthTransaction;
-use monad_eth_types::EthAddress;
+use monad_eth_types::{EthAddress, EthTransaction};
 use monad_executor::{Executor, ExecutorMetricsChain};
 use monad_executor_glue::{BlockSyncEvent, LedgerCommand, MonadEvent};
 use monad_state_backend::InMemoryState;
