+++ monad-eth-testutil/Cargo.toml
@@ -7,13 +7,8 @@ edition = "2021"
 bench = false
 
 [dependencies]
-monad-consensus-types = { path = "../monad-consensus-types" }
-monad-crypto = { path = "../monad-crypto" }
-monad-eth-block-policy = { path = "../monad-eth-block-policy" }
-monad-eth-tx = { path = "../monad-eth-tx" }
 monad-eth-types = { path = "../monad-eth-types" }
 monad-secp = { path = "../monad-secp" }
-monad-testutil = { path = "../monad-testutil" }
 monad-types = { path = "../monad-types" }
 rand = { workspace = true }
 reth-primitives = { workspace = true }
@@ -23,8 +18,12 @@ alloy-json-rpc = { workspace = true }
 alloy-primitives = { workspace = true }
 alloy-rlp = { workspace = true }
 alloy-rpc-client = { workspace = true }
+alloy-sol-macro = { workspace = true }
+alloy-sol-types = { workspace = true }
 alloy-transport = { workspace = true }
 clap = { workspace = true, features = ["derive"] }
+dashmap = { workspace = true }
+eyre = { workspace = true }
 futures = { workspace = true }
 itertools = { workspace = true }
 rand = { workspace = true }
@@ -37,7 +36,3 @@ tokio = { workspace = true, features = ["macros", "rt", "rt-multi-thread"] }
 tracing = { workspace = true, features = ["log", "std"] }
 tracing-subscriber = { workspace = true, features = ["env-filter", "std"] }
 url = { workspace = true }
-alloy-sol-macro = { workspace = true }
-alloy-sol-types = { workspace = true }
-dashmap = { workspace = true }
-eyre = { workspace = true }
