--- monad-eth-tx/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "monad-eth-tx"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-# this is necessary for criterion bench options
-# https://bheisler.github.io/criterion.rs/book/faq.html#cargo-bench-gives-unrecognized-option-errors-for-valid-command-line-options
-bench = false
-
-[dependencies]
-alloy-primitives = { workspace = true }
-alloy-rlp = { workspace = true }
-bytes = { workspace = true }
-reth-primitives = { workspace = true }
