+++ monad-eth-txpool/benches/common/controller.rs
@@ -1,18 +1,21 @@
+#[path = "../../test-utils/mod.rs"]
+mod test_utils;
+
 use bytes::Bytes;
 use itertools::Itertools;
 use monad_consensus_types::txpool::TxPool;
 use monad_crypto::NopSignature;
 use monad_eth_block_policy::{EthBlockPolicy, EthValidatedBlock};
-use monad_eth_testutil::{generate_block_with_txs, make_tx};
-use monad_eth_tx::EthSignedTransaction;
+use monad_eth_testutil::make_tx;
 use monad_eth_txpool::EthTxPool;
-use monad_eth_types::{Balance, EthAddress};
+use monad_eth_types::{Balance, EthAddress, EthSignedTransaction};
 use monad_state_backend::{InMemoryBlockState, InMemoryState, InMemoryStateInner};
 use monad_testutil::signing::MockSignatures;
 use monad_types::{Round, SeqNum};
 use rand::{seq::SliceRandom, Rng, SeedableRng};
 use rand_chacha::ChaCha8Rng;
 use reth_primitives::B256;
+use test_utils::generate_block_with_txs;
 
 const TRANSACTION_SIZE_BYTES: usize = 400;
