+++ monad-mock-swarm/Cargo.toml
@@ -43,7 +43,6 @@ monad-eth-block-policy = { path = "../monad-eth-block-policy" }
 monad-eth-block-validator = { path = "../monad-eth-block-validator" }
 monad-eth-ledger = { path = "../monad-eth-ledger" }
 monad-eth-testutil = { path = "../monad-eth-testutil" }
-monad-eth-tx = { path = "../monad-eth-tx" }
 monad-eth-txpool = { path = "../monad-eth-txpool" }
 monad-secp = { path = "../monad-secp" }
 monad-testutil = { path = "../monad-testutil" }
